Move `Traversals` helper methods elsewhere to be better encapsulated
