+++ mindmaps-graql/src/main/java/io/mindmaps/graql/internal/gremlin/fragment/InPlaysRoleFragment.java
@@ -4,8 +4,8 @@ import io.mindmaps.graql.internal.gremlin.FragmentPriority;
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversal;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
 
-import static io.mindmaps.graql.internal.gremlin.Traversals.inSubs;
-import static io.mindmaps.graql.internal.gremlin.Traversals.outSubs;
+import static io.mindmaps.graql.internal.gremlin.fragment.Fragments.inSubs;
+import static io.mindmaps.graql.internal.gremlin.fragment.Fragments.outSubs;
 import static io.mindmaps.util.Schema.EdgeLabel.PLAYS_ROLE;
 
 class InPlaysRoleFragment extends AbstractFragment {
