+++ mindmaps-graql/src/main/java/io/mindmaps/graql/internal/gremlin/fragment/InSubFragment.java
@@ -1,7 +1,6 @@
 package io.mindmaps.graql.internal.gremlin.fragment;
 
 import io.mindmaps.graql.internal.gremlin.FragmentPriority;
-import io.mindmaps.graql.internal.gremlin.Traversals;
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversal;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
 
@@ -13,7 +12,7 @@ class InSubFragment extends AbstractFragment {
 
     @Override
     public void applyTraversal(GraphTraversal<Vertex, Vertex> traversal) {
-        Traversals.inSubs(traversal);
+        Fragments.inSubs(traversal);
     }
 
     @Override
